Support Superday: improve user identification for Customer.io integration
Adds a fallback method to use the event user ID in case email is not available.

This method still prefers and uses email address whenever available. In case the email is required but not available, it will then prefer using the UUID of the user identified by the event instead.

**This is untested but submitted as part of the Support Superday activity**